+++ src/main/java/dev/chinh/portfolio/apps/wallet/dto/CreateRecurringRequest.java
@@ -1,11 +1,12 @@
 package dev.chinh.portfolio.apps.wallet.dto;
 
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public record CreateRecurringRequest(
     @NotNull Long walletId,
     @NotNull Long categoryId,
-    @NotNull @Positive double amount,
+    @NotNull @Positive BigDecimal amount,
     @NotBlank String type, // EXPENSE or INCOME
     @NotBlank String frequency, // DAILY, WEEKLY, MONTHLY, YEARLY
     @NotBlank String startDate, // YYYY-MM-DD
