+++ src/main/java/dev/chinh/portfolio/apps/wallet/dto/CreateTransactionRequest.java
@@ -1,12 +1,13 @@
 package dev.chinh.portfolio.apps.wallet.dto;
 
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public record CreateTransactionRequest(
     @NotNull Long walletId,
     Long categoryId,
     Long groupId,
-    @NotNull @DecimalMin("0.01") Double amount,
+    @NotNull @DecimalMin("0.01") BigDecimal amount,
     @NotBlank String type,
     String txnType,
     String note,
